[ASV-1350] Removed unnecessary code. Extracted methods to prevent the reuse of code.
