+++ app/src/vanilla/java/cm/aptoide/pt/presenter/LoginSignupCredentialsFlavorPresenter.java
@@ -8,13 +8,10 @@ import cm.aptoide.pt.crashreports.CrashReport;
 import cm.aptoide.pt.view.ThrowableToStringMapper;
 import java.util.Collection;
 import rx.Observable;
-import rx.android.schedulers.AndroidSchedulers;
 
 public class LoginSignupCredentialsFlavorPresenter extends LoginSignUpCredentialsPresenter {
 
   private final LoginSignUpCredentialsView view;
-  private final AccountAnalytics accountAnalytics;
-  private final AptoideAccountManager accountManager;
   private final ThrowableToStringMapper errorMapper;
   private final CrashReport crashReport;
   private final AccountNavigator accountNavigator;
@@ -28,8 +25,6 @@ public class LoginSignupCredentialsFlavorPresenter extends LoginSignUpCredential
     super(view, accountManager, crashReport, dismissToNavigateToMainView, navigateToHome,
         accountNavigator, permissions, requiredPermissions, errorMapper, accountAnalytics);
     this.view = view;
-    this.accountAnalytics = accountAnalytics;
-    this.accountManager = accountManager;
     this.errorMapper = errorMapper;
     this.crashReport = crashReport;
     this.accountNavigator = accountNavigator;
@@ -43,8 +38,13 @@ public class LoginSignupCredentialsFlavorPresenter extends LoginSignUpCredential
     handleAptoideShowSignUpEvent();
     handleClickOnTermsAndConditions();
     handleClickOnPrivacyPolicy();
-    handleAptoideSignUpEvent();
     showTCandPP();
+  }
+
+  protected Observable<AptoideCredentials> getAptoideSignUpEvent() {
+    return view.aptoideSignUpEvent()
+        .doOnNext(credentials -> showNotCheckedMessage(credentials.isChecked()))
+        .filter(AptoideCredentials::isChecked);
   }
 
   private void handleClickOnTermsAndConditions() {
@@ -65,41 +65,6 @@ public class LoginSignupCredentialsFlavorPresenter extends LoginSignUpCredential
         .compose(view.bindUntilEvent(View.LifecycleEvent.DESTROY))
         .subscribe(__ -> {
         }, err -> crashReport.log(err));
-  }
-
-  private void handleAptoideSignUpEvent() {
-    view.getLifecycleEvent()
-        .filter(event -> event.equals(View.LifecycleEvent.CREATE))
-        .flatMap(__ -> view.aptoideSignUpEvent()
-            .doOnNext(credentials -> showNotCheckedMessage(credentials.isChecked()))
-            .filter(AptoideCredentials::isChecked)
-            .doOnNext(click -> {
-              view.hideKeyboard();
-              view.showLoading();
-              lockScreenRotation();
-              accountAnalytics.sendAptoideSignUpButtonPressed();
-            })
-            .flatMapCompletable(
-                credentials -> accountManager.signUp(AptoideAccountManager.APTOIDE_SIGN_UP_TYPE,
-                    credentials)
-                    .observeOn(AndroidSchedulers.mainThread())
-                    .doOnCompleted(() -> {
-                      accountAnalytics.loginSuccess();
-                      navigateToCreateProfile();
-                      unlockScreenRotation();
-                      view.hideLoading();
-                    })
-                    .doOnError(throwable -> {
-                      accountAnalytics.sendSignUpErrorEvent(AccountAnalytics.LoginMethod.APTOIDE,
-                          throwable);
-                      view.showError(errorMapper.map(throwable));
-                      crashReport.log(throwable);
-                      unlockScreenRotation();
-                      view.hideLoading();
-                    }))
-            .retry())
-        .compose(view.bindUntilEvent(View.LifecycleEvent.DESTROY))
-        .subscribe();
   }
 
   private void showTCandPP() {
